requirements.txt
django
twisted
requests
django-progressive-web-app
huey
redis
pytest-django